save user

+++ internal/domain/user/model.go
@@ -4,7 +4,7 @@ import "github.com/nyakovchuk/vsch_church_bot/internal/domain/coordinates/model"
 
 type User struct {
 	ID          int
-	TgID        int
+	TgID        int64
 	Coordinates model.Coordinates
 	LangId      *int
 	Radius      int
